feat: :sparkles: Add percentage of determination/addiction/cooling time
Add percentage of determination/addiction/cooling time
