+++ logic/Client/MainWindow.xaml.cs
@@ -19,6 +19,7 @@ using Protobuf;
 using Playback;
 using CommandLine;
 using Preparation.Utility;
+using Preparation.Interface;
 
 // 目前MainWindow还未复现的功能：
 // private void ClickToSetMode(object sender, RoutedEventArgs e)
@@ -408,10 +409,22 @@ namespace Client
                                     switch (obj.MessageOfObjCase)
                                     {
                                         case MessageOfObj.MessageOfObjOneofCase.StudentMessage:
-
                                             if (humanOrButcher && obj.StudentMessage.PlayerId == playerID)
                                             {
                                                 human = obj.StudentMessage;
+                                            }
+                                            if (obj.StudentMessage.PlayerId < GameData.numOfStudent)
+                                            {
+                                                IStudentType occupation = (IStudentType)OccupationFactory.FindIOccupation(Transformation.ToStudentType(obj.StudentMessage.StudentType));
+                                                totalLife[obj.StudentMessage.PlayerId] = occupation.MaxHp;
+                                                totalDeath[obj.StudentMessage.PlayerId] = occupation.MaxGamingAddiction;
+                                                int i = 0;
+                                                foreach (var skill in occupation.ListOfIActiveSkill)
+                                                {
+                                                    var iActiveSkill = SkillFactory.FindIActiveSkill(skill);
+                                                    coolTime[i, obj.StudentMessage.PlayerId] = iActiveSkill.SkillCD;
+                                                    ++i;
+                                                }
                                             }
                                             listOfHuman.Add(obj.StudentMessage);
                                             break;
@@ -419,6 +432,14 @@ namespace Client
                                             if (!humanOrButcher && obj.TrickerMessage.PlayerId == playerID)
                                             {
                                                 butcher = obj.TrickerMessage;
+                                            }
+                                            IGhostType occupation1 = (IGhostType)OccupationFactory.FindIOccupation(Transformation.ToTrickerType(obj.TrickerMessage.TrickerType));
+                                            int j = 0;
+                                            foreach (var skill in occupation1.ListOfIActiveSkill)
+                                            {
+                                                var iActiveSkill = SkillFactory.FindIActiveSkill(skill);
+                                                coolTime[j, 4] = iActiveSkill.SkillCD;
+                                                ++j;
                                             }
                                             listOfButcher.Add(obj.TrickerMessage);
                                             break;
@@ -644,7 +665,7 @@ namespace Client
             return true;
         }
 
-        private void Refresh(object? sender, EventArgs e) // 已按照3.5版proto更新信息，circumstance栏未更新 log未更新
+        private void Refresh(object? sender, EventArgs e) //log未更新
         {
             // Bonus();
             if (WindowState == WindowState.Maximized)
@@ -652,16 +673,21 @@ namespace Client
             else
                 MaxButton.Content = "🗖";
             if (StatusBarsOfSurvivor != null)
-                for (int i = 4; i < 8; i++)
+            {
+                for (int i = 0; i < GameData.numOfStudent; i++)
                 {
-                    StatusBarsOfSurvivor[i - 4].SetFontSize(12 * UpperLayerOfMap.ActualHeight / 650);
+                    StatusBarsOfSurvivor[i].SetFontSize(12 * UpperLayerOfMap.ActualHeight / 650);
+                    StatusBarsOfSurvivor[i].NewData(totalLife, totalDeath, coolTime);
                 }
+            }
             if (StatusBarsOfHunter != null)
+            {
                 StatusBarsOfHunter.SetFontSize(12 * UpperLayerOfMap.ActualHeight / 650);
+                StatusBarsOfHunter.NewData(totalLife, totalDeath, coolTime);
+            }
             if (StatusBarsOfCircumstance != null)
                 StatusBarsOfCircumstance.SetFontSize(12 * UpperLayerOfMap.ActualHeight / 650);
             // 完成窗口信息更新
-
             if (!isClientStocked)
             {
                 unit = Math.Sqrt(UpperLayerOfMap.ActualHeight * UpperLayerOfMap.ActualWidth) / 50;
@@ -1362,10 +1388,10 @@ namespace Client
             { 6, 6, 0, 0, 7, 7, 6, 7, 7, 0, 0, 0, 0, 0, 0, 0, 14, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0, 0, 0, 0, 6 },
             { 6, 6, 15, 0, 0, 0, 7, 0, 0, 0, 0, 6, 6, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 11, 6, 0, 0, 0, 0, 0, 6 },
             { 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 6,6, 6, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6 },
-            { 6, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 15, 0, 7, 7, 7, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6 },
-            { 6, 0, 0, 0, 0, 0, 0, 0, 6, 6, 0, 0,8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 6 },
+            { 6, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 15, 0, 7, 7, 7, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6 },
+            { 6, 0, 0, 0, 0, 0, 0, 0, 6, 7, 0, 0,8, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 7, 7, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 6 },
             { 6, 0, 0, 0, 0, 0, 0, 0, 6, 7, 7, 0, 0, 0, 6, 6, 6, 11, 6, 0, 0, 6, 6, 6, 7, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 6 },
-            { 6, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 6, 6, 0, 0, 0, 0, 6, 0, 6, 7, 7, 6, 7, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 14, 6, 6, 6, 0, 0, 0, 0, 0, 7, 0, 0, 6, 0, 6 },
+            { 6, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 6, 6, 6, 6, 0, 0, 0, 6, 0, 6, 7, 7, 6, 7, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 14, 6, 6, 6, 0, 0, 0, 0, 0, 7, 0, 0, 6, 0, 6 },
             { 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 6, 7, 0, 7, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 7, 6, 0, 6, 6, 0, 6 },
             { 6, 0, 0, 0, 0, 0, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6, 7, 0, 0, 7, 6, 0, 0, 0, 0, 0, 0, 0, 0, 8, 0, 0, 0, 0, 0, 6, 0, 0, 0, 0, 7, 6, 6, 6, 0, 0, 6 },
             { 6, 0, 0, 0, 0, 0, 6, 7, 7, 0, 0, 0, 0, 0, 0, 0, 0, 0, 7, 0, 0, 0, 7, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 11, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 6 },
@@ -1382,5 +1408,7 @@ namespace Client
         bool isEmergencyOpened = false;
         bool isEmergencyDrawed = false;
         const double radiusTimes = 1.0 * Preparation.Utility.GameData.characterRadius / Preparation.Utility.GameData.numOfPosGridPerCell;
+        private int[] totalLife = new int[4] { 100, 100, 100, 100 }, totalDeath = new int[4] { 100, 100, 100, 100 };
+        private int[,] coolTime = new int[3, 5] { { 100, 100, 100, 100, 100 }, { 100, 100, 100, 100, 100 }, { 100, 100, 100, 100, 100 } };
     }
 }
