+++ logic/Preparation/Preparation.csproj
@@ -7,5 +7,13 @@
     <StartupObject />
     <Nullable>enable</Nullable>
   </PropertyGroup>
+	
+	<ItemGroup>
+		<PackageReference Include="Google.Protobuf" Version="3.22.1" />
+	</ItemGroup>
 
+	<ItemGroup>
+		<ProjectReference Include="..\..\dependency\proto\Protos.csproj" />
+	</ItemGroup>
+	
 </Project>
